+++ openeogeotrellis/backend.py
@@ -1,3 +1,4 @@
+import abc
 import datetime as dt
 import io
 import json
@@ -97,6 +98,7 @@ from openeogeotrellis.integrations.kubernetes import (
 )
 from openeogeotrellis.integrations.stac import ResilientStacIO
 from openeogeotrellis.integrations.traefik import Traefik
+from openeogeotrellis.job_options import JobOptions, K8SOptions
 from openeogeotrellis.job_registry import (
     DoubleJobRegistry,
     ZkJobRegistry,
@@ -594,6 +596,9 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
                 },
             }
         }
+
+    def processing_parameters(self) -> list:
+        return JobOptions.list_options()
 
     def load_disk_data(
             self, format: str, glob_pattern: str, options: dict, load_params: LoadParameters, env: EvalEnv
@@ -1126,6 +1131,7 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
             elif "outofmemoryerror" in root_cause_class_name.lower():
                 summary = "Your batch job failed because the 'driver' used too much java memory. Consider increasing driver-memory or contact the developers to investigate."
             elif is_spark_exception:
+                retry_message = "Simply try submitting again, or use batch job logs to find more detailed information in case of persistent failures. Increasing executor memory may help if the root cause is not clear from the logs."
                 if missing_sentinel1_band:
                     summary = (f"Requested band '{missing_sentinel1_band}' is not present in Sentinel 1 tile;"
                                f' try specifying a "polarization" property filter according to the table at'
@@ -1177,11 +1183,14 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
                         )
                         or "has failed the maximum allowable number of times" in root_cause_message
                     ):
-                        summary = f"A part of your process graph failed multiple times. Simply try submitting again, or use batch job logs to find more detailed information in case of persistent failures. Increasing executor memory may help if the root cause is not clear from the logs."
+                        summary = f"A part of your process graph failed multiple times. " + retry_message
                     else:
                         summary = f"Exception during Spark execution: {root_cause_class_name}: {root_cause_message}"
                 else:
-                    summary = f"Exception during Spark execution: {root_cause_class_name}"
+                    if root_cause_class_name == "java.io.EOFException":
+                        summary = f"Exception during Spark execution. " + retry_message
+                    else:
+                        summary = f"Exception during Spark execution: {root_cause_class_name}"
             else:
                 summary = f"{root_cause_class_name}: {root_cause_message}"
             summary = str_truncate(summary, width=width)
@@ -1233,7 +1242,15 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
             start_index = full_stacktrace.rfind("\n", 0, needle_index) + 1
             if start_index == -1:
                 start_index = 0
-            udf_stacktrace = full_stacktrace[start_index:].rstrip()
+
+            # No need for the Java stack trace if we found UDF stack trace.
+            # Triggering an exception in the UDF code with
+            #   jvm.org.openeo.dumpGeoJson("", jvm.scala.Some(""))
+            # will not add a jvm stacktrace. Instead, it wil give a connection lost error.
+            end_index = full_stacktrace.find("\n\tat ", start_index)
+            if end_index == -1:
+                end_index = len(full_stacktrace)
+            udf_stacktrace = full_stacktrace[start_index:end_index].rstrip()
             udf_stacktrace = GeoPySparkBackendImplementation.collapse_stack_strace(udf_stacktrace, width)
             return udf_stacktrace
         return None
@@ -1429,6 +1446,7 @@ def get_elastic_job_registry(
         api_url=config.ejr_api,
         backend_id=config.ejr_backend_id,
         session=requests_session,
+        preserialize_process=config.ejr_preserialize_process,
     )
     # Get credentials from env (preferably) or vault (as fallback).
     ejr_creds = get_ejr_credentials_from_env(strict=False)
@@ -1944,24 +1962,7 @@ class GpsBatchJobs(backend.BatchJobs):
 
             raise OpenEOApiException(f"invalid value {value} for job_option {job_option_key}")
 
-        def as_max_soft_errors_ratio_arg() -> str:
-            value = job_options.get("soft-errors")
 
-            if value == "false":
-                return "0.0"
-            elif value == None:
-                return str(get_backend_config().default_soft_errors)
-            elif value == "true":
-                return "1.0"
-            elif isinstance(value, bool):
-                return "1.0" if value else "0.0"
-            elif isinstance(value, (int, float)) and 0.0 <= value <= 1.0:
-                return str(value)
-
-            raise OpenEOApiException(message=f"invalid value {value} for job_option soft-errors; "
-                                             f"supported values include false/true and values in the "
-                                             f"interval [0.0, 1.0]",
-                                     status_code=400)
 
         def as_logging_threshold_arg() -> str:
             if JOB_OPTION_LOGGING_THRESHOLD in job_options:
@@ -1986,55 +1987,25 @@ class GpsBatchJobs(backend.BatchJobs):
             )
 
         isKube = ConfigParams().is_kube_deploy
-        driver_memory = job_options.get("driver-memory", get_backend_config().default_driver_memory )
-        driver_memory_overhead = job_options.get("driver-memoryOverhead", get_backend_config().default_driver_memoryOverhead)
-        executor_memory = job_options.get("executor-memory", get_backend_config().default_executor_memory)
-        executor_memory_overhead = job_options.get("executor-memoryOverhead", get_backend_config().default_executor_memoryOverhead )
-        driver_cores = str(job_options.get("driver-cores", 1 if isKube else 5))
-        executor_cores = str(job_options.get("executor-cores", 1 if isKube else 2))
-        executor_corerequest = job_options.get("executor-request-cores", "NONE")
-        if executor_corerequest == "NONE":
-            executor_corerequest = str(int(executor_cores)/2*1000)+"m"
-        max_executors = str(job_options.get("max-executors", get_backend_config().default_max_executors))
-        executor_threads_jvm = str(job_options.get("executor-threads-jvm", get_backend_config().default_executor_threads_jvm))
-        gdal_dataset_cache_size = str(job_options.get("gdal-dataset-cache-size", get_backend_config().default_gdal_dataset_cache_size))
-        gdal_cachemax = str(job_options.get("gdal-cachemax", get_backend_config().default_gdal_cachemax ))
+
+        if isKube:
+            options = K8SOptions.from_dict(job_options)
+        else:
+            options = JobOptions.from_dict(job_options)
+
+        executor_memory_overhead = options.executor_memory_overhead
+
         queue = job_options.get("queue", "default")
         profile = as_boolean_arg("profile", default_value="false")
-        max_soft_errors_ratio = as_max_soft_errors_ratio_arg()
-        task_cpus = str(job_options.get("task-cpus", 1))
-        archives = ",".join(job_options.get("udf-dependency-archives", []))
-        py_files = job_options.get("udf-dependency-files", [])
-        use_goofys = as_boolean_arg("goofys", default_value="false") != "false"
-        mount_tmp = as_boolean_arg("mount_tmp", default_value="false") != "false"
-        use_pvc = as_boolean_arg("spark_pvc", default_value="false") != "false"
+
         logging_threshold = as_logging_threshold_arg()
         propagatable_web_app_driver_envars = {
             envar: os.environ.get(envar, "")
             for envar in os.environ.get("OPENEO_PROPAGATABLE_WEB_APP_DRIVER_ENVARS", "").split()
         }
 
-        as_bytes = self._jvm.org.apache.spark.util.Utils.byteStringAsBytes
+        options.validate()
 
-        if as_bytes(executor_memory) + as_bytes(executor_memory_overhead) > as_bytes(get_backend_config().max_executor_or_driver_memory):
-            raise OpenEOApiException(
-                message=f"Requested too much executor memory: {executor_memory} + {executor_memory_overhead}, the max for this instance is: {get_backend_config().max_executor_or_driver_memory}",
-                status_code=400)
-        if as_bytes(driver_memory) + as_bytes(driver_memory_overhead) > as_bytes(get_backend_config().max_executor_or_driver_memory):
-            raise OpenEOApiException(
-                message=f"Requested too much driver memory: {driver_memory} + {driver_memory_overhead}, the max for this instance is: {get_backend_config().max_executor_or_driver_memory}",
-                status_code=400)
-
-        if isKube:
-            max_cores = 4
-            if int(executor_cores) > max_cores:
-                raise OpenEOApiException(
-                    message=f"Requested too many executor cores: {executor_cores} , the max for this instance is: {max_cores}",
-                    status_code=400)
-            if int(driver_cores) > max_cores:
-                raise OpenEOApiException(
-                    message=f"Requested too many driver cores: {driver_cores} , the max for this instance is: {max_cores}",
-                    status_code=400)
 
 
         def serialize_dependencies() -> str:
@@ -2058,12 +2029,13 @@ class GpsBatchJobs(backend.BatchJobs):
         if get_backend_config().setup_kerberos_auth:
             setup_kerberos_auth(self._principal, self._key_tab, self._jvm)
 
-        memOverheadBytes = as_bytes(executor_memory_overhead)
+        as_bytes = self._jvm.org.apache.spark.util.Utils.byteStringAsBytes
+        memOverheadBytes = as_bytes(options.executor_memory_overhead)
         jvmOverheadBytes = as_bytes("128m")
 
         # By default, Python uses the space reserved by `spark.executor.memoryOverhead` but no limit is enforced.
         # When `spark.executor.pyspark.memory` is specified, Python will only use this memory and no more.
-        python_max = job_options.get("python-memory", get_backend_config().default_python_memory)
+        python_max = options.python_memory
         if python_max is not None:
             python_max = as_bytes(python_max)
             if "executor-memoryOverhead" not in job_options:
@@ -2078,21 +2050,37 @@ class GpsBatchJobs(backend.BatchJobs):
                 python_max = -1
             executor_memory_overhead = f"{memOverheadBytes // (1024 ** 2)}m"
 
-        if as_bytes(executor_memory) + as_bytes(executor_memory_overhead) + python_max > as_bytes(
+        if as_bytes(options.executor_memory) + as_bytes(executor_memory_overhead) + python_max > as_bytes(
                 get_backend_config().max_executor_or_driver_memory
         ):
             raise OpenEOApiException(
                 message=f"Requested too much executor memory: "
-                        + f"{executor_memory} + {executor_memory_overhead} + {python_max // (1024 ** 2)}m, "
+                        + f"{options.executor_memory} + {executor_memory_overhead} + {python_max // (1024 ** 2)}m, "
                         + f"the max for this instance is: {get_backend_config().max_executor_or_driver_memory}",
                 status_code=400,
             )
+
+        user_provided_jar_path = None
+        if "openeo-jar-path" in job_options and (
+                job_options["openeo-jar-path"].startswith("https://artifactory.vgt.vito.be/artifactory/libs-release-public/org/openeo/geotrellis-extensions")
+                or job_options["openeo-jar-path"].startswith("https://artifactory.vgt.vito.be/artifactory/libs-snapshot-public/org/openeo/geotrellis-extensions/") ) :
+            user_provided_jar_path = job_options["openeo-jar-path"]
 
         if isKube:
             # TODO: get rid of this "isKube" anti-pattern, it makes testing of this whole code path practically impossible
 
             # TODO: eliminate these local imports
             from kubernetes.client.rest import ApiException
+
+            use_goofys = as_boolean_arg("goofys", default_value="false") != "false"
+            mount_tmp = as_boolean_arg("mount_tmp", default_value="false") != "false"
+            use_pvc = as_boolean_arg("spark_pvc", default_value="false") != "false"
+
+            executor_corerequest = job_options.get("executor-request-cores", "NONE")
+            if executor_corerequest == "NONE":
+                executor_corerequest = str(int(options.executor_cores) / 2 * 1000) + "m"
+
+            gdal_cachemax = str(job_options.get("gdal-cachemax", get_backend_config().default_gdal_cachemax))
 
 
             api_instance_custom_object = kube_client("CustomObject")
@@ -2183,23 +2171,23 @@ class GpsBatchJobs(backend.BatchJobs):
                 metadata_file=JOB_METADATA_FILENAME,
                 job_id_short=truncate_job_id_k8s(job_id),
                 job_id_full=job_id,
-                driver_cores=driver_cores,
-                driver_memory=driver_memory,
-                driver_memory_overhead=driver_memory_overhead,
-                executor_cores=executor_cores,
+                driver_cores=str(options.driver_cores),
+                driver_memory=options.driver_memory,
+                driver_memory_overhead=options.driver_memory_overhead,
+                executor_cores=options.executor_cores,
                 executor_corerequest=executor_corerequest,
-                executor_memory=executor_memory,
-                executor_memory_overhead=executor_memory_overhead,
-                executor_threads_jvm=executor_threads_jvm,
+                executor_memory=options.executor_memory,
+                executor_memory_overhead=options.executor_memory_overhead,
+                executor_threads_jvm=str(options.executor_threads_jvm),
                 python_max_memory = python_max,
-                max_executors=max_executors,
+                max_executors=options.max_executors,
                 api_version=api_version,
                 dependencies="[]",  # TODO: use `serialize_dependencies()` here instead? It's probably messy to get that JSON string correctly encoded in the rendered YAML.
                 user_id=user_id,
                 user_id_short=truncate_user_id_k8s(user_id),
-                max_soft_errors_ratio=max_soft_errors_ratio,
-                task_cpus=task_cpus,
-                gdal_dataset_cache_size=gdal_dataset_cache_size,
+                max_soft_errors_ratio=options.soft_errors_arg(),
+                task_cpus=options.task_cpus,
+                gdal_dataset_cache_size=str(options.gdal_dataset_cache_size),
                 gdal_cachemax=gdal_cachemax,
                 current_time=int(time.time()),
                 aws_access_key_id=os.environ.get("AWS_ACCESS_KEY_ID"),
@@ -2215,8 +2203,8 @@ class GpsBatchJobs(backend.BatchJobs):
                 swift_bucket=bucket,
                 zookeeper_nodes=os.environ.get("ZOOKEEPERNODES"),
                 eodata_mount=eodata_mount,
-                archives=archives,
-                py_files = py_files,
+                archives=",".join(options.udf_dependency_archives),
+                py_files = options.udf_dependency_files,
                 logging_threshold=logging_threshold,
                 mount_tmp=mount_tmp,
                 use_pvc=use_pvc,
@@ -2237,7 +2225,8 @@ class GpsBatchJobs(backend.BatchJobs):
                 propagatable_web_app_driver_envars=propagatable_web_app_driver_envars,
                 provide_s3_profiles_and_tokens=get_backend_config().provide_s3_profiles_and_tokens,
                 batch_job_cfg_secret_name=batch_job_cfg_secret_name,
-                batch_job_config_dir=get_backend_config().batch_job_config_dir
+                batch_job_config_dir=get_backend_config().batch_job_config_dir,
+                openeo_jar_path=options.openeo_jar_path or 'local:///opt/geotrellis-extensions-static.jar'
             )
 
             with self._double_job_registry as dbl_registry:
@@ -2312,15 +2301,14 @@ class GpsBatchJobs(backend.BatchJobs):
                     dbl_registry.set_status(job_id=job_id, user_id=user_id, status=JOB_STATUS.ERROR)
 
         else:
-            ########"
             submit_script = "submit_batch_job_spark3.sh"
             script_location = pkg_resources.resource_filename("openeogeotrellis.deploy", submit_script)
 
             image_name = job_options.get("image-name", os.environ.get("YARN_CONTAINER_RUNTIME_DOCKER_IMAGE"))
 
             extra_py_files=""
-            if len(py_files)>0:
-                extra_py_files = "," + py_files.join(",")
+            if options.udf_dependency_files is not None and len(options.udf_dependency_files)>0:
+                extra_py_files = ",".join(options.udf_dependency_files)
 
             job_work_dir = self.get_job_work_dir(job_id=job_id)
 
@@ -2347,7 +2335,7 @@ class GpsBatchJobs(backend.BatchJobs):
                 temp_properties_file.flush()
 
                 job_name = "openEO batch_{title}_{j}_user {u}".format(title=job_title, j=job_id, u=user_id)
-                args = [
+                args: list[str] = [
                     script_location,
                     job_name,
                     job_specification_file.name,
@@ -2371,24 +2359,24 @@ class GpsBatchJobs(backend.BatchJobs):
                 else:
                     args.append("0.4.0")
 
-                args.append(driver_memory)
-                args.append(executor_memory)
-                args.append(executor_memory_overhead)
-                args.append(driver_cores)
-                args.append(executor_cores)
-                args.append(driver_memory_overhead)
+                args.append(options.driver_memory)
+                args.append(options.executor_memory)
+                args.append(options.executor_memory_overhead)
+                args.append(str(options.driver_cores))
+                args.append(str(options.executor_cores))
+                args.append(options.driver_memory_overhead)
                 args.append(queue)
                 args.append(profile)
                 args.append(serialize_dependencies())
                 args.append(self.get_submit_py_files()+extra_py_files)
-                args.append(max_executors)
+                args.append(str(options.max_executors))
                 args.append(user_id)
                 args.append(job_id)
-                args.append(max_soft_errors_ratio)
-                args.append(task_cpus)
+                args.append(options.soft_errors_arg())
+                args.append(str(options.task_cpus))
                 args.append(sentinel_hub_client_alias)
                 args.append(temp_properties_file.name)
-                args.append(archives)
+                args.append(",".join(options.udf_dependency_archives))
                 args.append(logging_threshold)
                 args.append(os.environ.get(ConfigGetter.OPENEO_BACKEND_CONFIG, ""))
                 args.append(str(job_work_dir / UDF_PYTHON_DEPENDENCIES_FOLDER_NAME))
@@ -2412,8 +2400,8 @@ class GpsBatchJobs(backend.BatchJobs):
                     log.info(f"Submitting job with command {args!r}")
                     d = dict(**os.environ)
                     d["YARN_CONTAINER_RUNTIME_DOCKER_IMAGE"] = image_name
-                    if "openeo-jar-path" in job_options and job_options["openeo-jar-path"].startswith("https://artifactory.vgt.vito.be/artifactory/libs-release-public/org/openeo/geotrellis-extensions"):
-                        d["OPENEO_GEOTRELLIS_JAR"] = job_options["openeo-jar-path"]
+                    if options.openeo_jar_path is not None:
+                        d["OPENEO_GEOTRELLIS_JAR"] = options.openeo_jar_path
                     script_output = subprocess.check_output(args, stderr=subprocess.STDOUT, universal_newlines=True, env=d)
                     log.info(f"Submitted job, output was: {script_output}")
                 except CalledProcessError as e:
@@ -2424,7 +2412,6 @@ class GpsBatchJobs(backend.BatchJobs):
                 application_id = self._extract_application_id(script_output)
                 log.info("mapped job_id %s to application ID %s" % (job_id, application_id))
 
-                #########
                 with self._double_job_registry as dbl_registry:
                     dbl_registry.set_application_id(job_id=job_id, user_id=user_id, application_id=application_id)
                     dbl_registry.set_status(job_id=job_id, user_id=user_id, status=JOB_STATUS.QUEUED)
@@ -2899,7 +2886,18 @@ class GpsBatchJobs(backend.BatchJobs):
 
         job_dir = self.get_job_output_dir(job_id=job_id)
 
-        results_metadata = self.load_results_metadata(job_id, user_id)
+        results_metadata = None
+        try:
+            with self._double_job_registry as registry:
+                job_dict = registry.elastic_job_registry.get_job(job_id, user_id=user_id)
+                if "results_metadata" in job_dict:
+                    results_metadata = job_dict["results_metadata"]
+        except Exception as e:
+            logger.warning(
+                "Could not retrieve result metadata from job tracker %s", e, exc_info=True, extra={"job_id": job_id}
+            )
+        if results_metadata is None or len(results_metadata) == 0:
+            results_metadata = self.load_results_metadata(job_id, user_id)
         out_assets = results_metadata.get("assets", {})
         out_metadata = out_assets.get("out", {})
         bands = [Band(*properties) for properties in out_metadata.get("bands", [])]
@@ -2977,6 +2975,7 @@ class GpsBatchJobs(backend.BatchJobs):
         """
         Reads the metadata json file from the job directory and returns it.
         """
+
         metadata_file = self.get_results_metadata_path(job_id=job_id)
 
         if ConfigParams().use_object_storage:
@@ -3235,43 +3234,6 @@ class GpsBatchJobs(backend.BatchJobs):
         if assembled_folders:
             logger.info("Deleted Sentinel Hub assembled folder(s) {fs} for batch job {j}"
                         .format(fs=assembled_folders, j=job_id), extra={'job_id': job_id})
-
-    def delete_jobs_before(
-        self,
-        upper: dt.datetime,
-        *,
-        user_ids: Optional[List[str]] = None,
-        dry_run: bool = True,
-        include_ongoing: bool = True,
-        include_done: bool = True,
-        user_limit: Optional[int] = 1000,
-    ) -> None:
-        # TODO #632 #863 #1123 #1165 remove this dead code path?
-        with self._double_job_registry as registry, TimingLogger(
-            title=f"Collecting jobs to delete: {upper=} {user_ids=} {include_ongoing=} {include_done=}",
-            logger=logger,
-        ):
-            jobs_before = registry.get_all_jobs_before(
-                upper,
-                user_ids=user_ids,
-                include_ongoing=include_ongoing,
-                include_done=include_done,
-                user_limit=user_limit,
-            )
-        logger.info(f"Collected {len(jobs_before)} jobs to delete")
-
-        with TimingLogger(title=f"Deleting {len(jobs_before)} jobs", logger=logger):
-
-            for job_info in jobs_before:
-                job_id = job_info["job_id"]
-                user_id = job_info["user_id"]
-                if not dry_run:
-                    logger.info(f"Deleting {job_id=} from {user_id=}")
-                    self._delete_job(
-                        job_id=job_id, user_id=user_id, propagate_errors=True
-                    )
-                else:
-                    logger.info(f"Dry run: not deleting {job_id=} from {user_id=}")