+++ openeogeotrellis/workspace/stac_api_workspace.py
@@ -54,12 +54,12 @@ class StacApiWorkspace(Workspace):
         self._asset_alternate_id = asset_alternate_id
 
     def import_file(self, common_path: Union[str, Path], file: Path, merge: str, remove_original: bool = False) -> str:
-        raise NotImplementedError
+        raise NotImplementedError(f"export_workspace: import_file to {self.root_url} not implemented {common_path}")
 
     def import_object(
         self, common_path: Union[str, Path], s3_uri: str, merge: str, remove_original: bool = False
     ) -> str:
-        raise NotImplementedError
+        raise NotImplementedError(f"export_workspace: import_file to {self.root_url} not implemented {common_path}")
 
     def merge(self, stac_resource: STACObject, target: PurePath, remove_original: bool = False) -> STACObject:
         self._assert_catalog_supports_necessary_api()
@@ -137,6 +137,9 @@ class StacApiWorkspace(Workspace):
         bare_collection = collection.clone()
         bare_collection.id = collection_id
         bare_collection.remove_hierarchical_links()
+        bare_collection.remove_links("queryables")
+        bare_collection.remove_links("aggregations")
+        bare_collection.remove_links("aggregate")
         bare_collection.extra_fields.update(self._additional_collection_properties)
 
         request_json = bare_collection.to_dict(include_self_link=False)