+++ tests/conftest.py
@@ -18,9 +18,11 @@ import moto
 import moto.server
 import openeo_driver
 from openeo_driver.config import OpenEoBackendConfig
+from openeo.testing.io import TestDataLoader
 from openeogeotrellis.integrations.identity import IDPTokenIssuer
 
 from openeogeotrellis.config.s3_config import AWSConfig
+from openeo_driver.integrations.s3.client import S3ClientBuilder
 
 import openeogeotrellis
 import pytest
@@ -264,6 +266,11 @@ class _Sleeper:
 
 
 @pytest.fixture
+def test_data() -> TestDataLoader:
+    return TestDataLoader(root=TEST_DATA_ROOT)
+
+
+@pytest.fixture
 def fast_sleep(time_machine) -> typing.Iterator[_Sleeper]:
     """
     Fixture using `time_machine` to make `sleep` instant and update the current time.
@@ -331,23 +338,15 @@ def config_overrides() -> dict:
 
 
 @pytest.fixture
-def _config_overrides(config_overrides, _dynamic_overrides) -> dict:
-    # use config_overrides to set manual overrides for your test
-    return {
-        **_dynamic_overrides,
-        **config_overrides
-    }
-
-
-@pytest.fixture
-def set_config_overrides(_config_overrides) -> typing.Generator[GpsBackendConfig, None, None]:
+def _set_config_overrides(config_overrides, _dynamic_overrides) -> typing.Iterator[GpsBackendConfig]:
+    all_config_overrides = {**_dynamic_overrides, **config_overrides}
     openeo_driver_overrides = {}
-    for config_key, config_override_value in _config_overrides.items():
+    for config_key, config_override_value in all_config_overrides.items():
         if hasattr(OpenEoBackendConfig, config_key):
             openeo_driver_overrides[config_key] = config_override_value
 
     geopyspark_driver_overrides = {}
-    for config_key, config_override_value in _config_overrides.items():
+    for config_key, config_override_value in all_config_overrides.items():
         if hasattr(OpenEoBackendConfig, config_key):
             openeo_driver_overrides[config_key] = config_override_value
         elif hasattr(openeogeotrellis.config.GpsBackendConfig, config_key):
@@ -360,7 +359,9 @@ def set_config_overrides(_config_overrides) -> typing.Generator[GpsBackendConfig
 
 
 @pytest.fixture
-def backend_implementation(batch_job_output_root, job_registry, set_config_overrides) -> "GeoPySparkBackendImplementation":
+def backend_implementation(
+    batch_job_output_root, job_registry, _set_config_overrides
+) -> "GeoPySparkBackendImplementation":
     from openeogeotrellis.backend import GeoPySparkBackendImplementation
 
     backend = GeoPySparkBackendImplementation(
@@ -450,7 +451,7 @@ def aws_credentials(monkeypatch):
 
 
 @pytest.fixture(scope="function")
-def mock_s3_resource(aws_credentials):
+def mock_s3_resource(aws_credentials, mock_s3_client):
     if moto_server_address is None:
         with moto.mock_aws():
             yield boto3.resource("s3", region_name=TEST_AWS_REGION_NAME)
@@ -458,12 +459,22 @@ def mock_s3_resource(aws_credentials):
         yield boto3.resource("s3", region_name=TEST_AWS_REGION_NAME, endpoint_url=moto_server_address)
 
 @pytest.fixture(scope="function")
-def mock_s3_client(aws_credentials):
+def mocked_s3_client(aws_credentials):
     if moto_server_address is None:
         with moto.mock_aws():
             yield boto3.client("s3", region_name=TEST_AWS_REGION_NAME)
     else:
         yield boto3.client("s3", region_name=TEST_AWS_REGION_NAME, endpoint_url=moto_server_address)
+
+
+@pytest.fixture(scope="function")
+def mock_s3_client(mocked_s3_client, monkeypatch):
+    def _get_client(*args, **kwargs):
+        return mocked_s3_client
+
+    # monkeypatch in case motoserver runs standalone
+    monkeypatch.setattr(S3ClientBuilder, "from_region", _get_client)
+    yield mocked_s3_client
 
 @pytest.fixture(scope="function")
 def mock_sts_client(monkeypatch):